+++ backend/.gitignore
@@ -1,0 +1,159 @@
+# -----------------------------------------------------------------------------
+# ENVIRONMENTS
+# -----------------------------------------------------------------------------
+
+ENV/
+env/
+env.bak/
+env_*/
+env3/
+venv/
+venv.bak/
+.env/
+.venv/
+
+# -----------------------------------------------------------------------------
+# BACKEND - DJANGO / PYTHON
+# -----------------------------------------------------------------------------
+
+# Byte-compiled / optimized / DLL files
+__pycache__/
+*.py[cod]
+*$py.class
+.pytest_cache/
+
+# C extensions
+*.so
+
+# Distribution / packaging
+.Python
+build/
+develop-eggs/
+dist/
+downloads/
+eggs/
+.eggs/
+lib/
+lib64/
+parts/
+sdist/
+var/
+wheels/
+share/python-wheels/
+*.egg-info/
+.installed.cfg
+*.egg
+MANIFEST
+
+# PyInstaller
+#  Usually these files are written by a python script from a template
+#  before PyInstaller builds the exe, so as to inject date/other infos into it.
+
+*.manifest
+*.spec
+
+# Installer logs
+pip-log.txt
+pip-delete-this-directory.txt
+
+# Unit test / coverage reports
+htmlcov
+locust
+.tox
+.nox
+coverage
+.coverage
+.coverage.*
+coverage.xml
+*.cover
+.cache
+nosetests.xml
+.hypothesis
+.pytest_cache
+.pytest-queries
+.scannerwork
+
+# Translations
+*.mo
+*.pot
+
+# Django stuff:
+local_settings.py
+local.py
+db.sqlite3
+media/
+*.box
+*.pyc
+*.pyo
+bower_components/
+configs/vault/
+deployment_*.py
+dump.rdb
+nohup.out
+octave-core
+parser.out
+parsetab.py
+personal_data/
+backups/
+mod_wsgi*/
+postgres-data
+redis-stable
+static/
+staticfiles/
+
+# Flask stuff:
+instance/
+.webassets-cache
+
+# Scrapy stuff:
+.scrapy
+
+# Sphinx documentation
+docs/build/
+docs/_build
+docs/_autosummary
+
+# PyBuilder
+target/
+
+# Jupyter Notebook
+.ipynb_checkpoints/
+*.ipynb
+
+# IPython
+profile_default/
+ipython_config.py
+
+# pyenv
+.python-version
+
+# celery beat schedule file
+celerybeat-schedule
+
+# SageMath parsed files
+*.sage.py
+
+# Spyder project settings
+.spyderproject
+.spyproject
+
+# Rope project settings
+.ropeproject
+
+# mkdocs documentation
+site
+
+# mypy
+mypyls/
+.mypy_cache/
+.dmypy.json
+dmypy.json
+
+# Pyre type checker
+.pyre/
+
+# pyenv
+.python-version
+
+# Physical quantities module (Valispace version)
+src/pint
